Improve Box + Stack components
