+++ packages/canon/src/components/Box/Docs.mdx
@@ -1,13 +1,7 @@
 import { Canvas, Meta, Unstyled } from '@storybook/blocks';
 import * as BoxStories from './Box.stories';
-import { Chip } from '../../../docs/components/Chip';
-import * as Table from '../../../docs/components/Table';
-import { Box } from './Box';
-import { Stack } from '../Stack';
 import { Title, Text } from '../../../docs/components';
 import { Padding } from './docs/padding';
-import { listResponsiveValues, listColorValues } from '../../utils/list-values';
-import { responsiveProperties } from './sprinkles.css';
 import { PropsTable } from './docs/props-table';
 import { SpacingTable } from './docs/spacing-table';
