+++ packages/canon/src/components/Box/index.tsx
@@ -15,4 +15,4 @@
  */
 export { Box } from './Box';
 export type * from './types';
-export { breakpoints, space, themes } from './properties';
+export { breakpoints, space, themes } from '../../layout/properties';
