+++ packages/canon/src/components/Box/types.ts
@@ -13,19 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { breakpoints, space, themes } from './properties';
+
+import { Breakpoint, SpaceProps, ColorProps } from '../../layout/types';
 
 /** @public */
-export type Breakpoint = keyof typeof breakpoints;
-
-/** @public */
-export type Space = keyof typeof space;
-
-/** @public */
-export type Theme = keyof typeof themes;
-
-/** @public */
-export type Display =
+export type DisplayProps =
   | 'flex'
   | 'none'
   | 'inline'
@@ -33,19 +25,19 @@ export type Display =
   | Partial<Record<Breakpoint, 'flex' | 'none' | 'inline' | 'block'>>;
 
 /** @public */
-export type FlexDirection =
+export type FlexDirectionProps =
   | 'row'
   | 'column'
   | Partial<Record<Breakpoint, 'row' | 'column'>>;
 
 /** @public */
-export type FlexWrap =
+export type FlexWrapProps =
   | 'wrap'
   | 'nowrap'
   | Partial<Record<Breakpoint, 'wrap' | 'nowrap'>>;
 
 /** @public */
-export type JustifyContent =
+export type JustifyContentProps =
   | 'stretch'
   | 'flex-start'
   | 'center'
@@ -65,7 +57,7 @@ export type JustifyContent =
     >;
 
 /** @public */
-export type AlignItems =
+export type AlignItemsProps =
   | 'stretch'
   | 'flex-start'
   | 'center'
@@ -75,7 +67,7 @@ export type AlignItems =
     >;
 
 /** @public */
-export type BorderRadius =
+export type BorderRadiusProps =
   | 'none'
   | 'small'
   | 'medium'
@@ -83,94 +75,15 @@ export type BorderRadius =
   | Partial<Record<Breakpoint, 'none' | 'small' | 'medium' | 'full'>>;
 
 /** @public */
-export type Gap = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type PaddingLeft = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type PaddingRight = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type PaddingTop = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type PaddingBottom = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type Padding = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type PaddingX = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type PaddingY = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type MarginLeft = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type MarginRight = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type MarginTop = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type MarginBottom = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type Margin = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type MarginX = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type MarginY = Space | Partial<Record<Breakpoint, Space>>;
-
-/** @public */
-export type Background =
-  | 'background'
-  | 'elevation1'
-  | 'elevation2'
-  | 'transparent'
-  | Partial<
-      Record<Theme, 'background' | 'elevation1' | 'elevation2' | 'transparent'>
-    >;
-
-/** @public */
-export type Color =
-  | 'primary'
-  | 'secondary'
-  | 'error'
-  | Partial<Record<Theme, 'primary' | 'secondary' | 'error'>>;
-
-/** @public */
-export interface BoxProps {
+export interface BoxProps extends SpaceProps, ColorProps {
   as?: keyof JSX.IntrinsicElements;
-  background?: Background;
   children?: React.ReactNode;
-  color?: Color;
-  display?: Display;
-  flexDirection?: FlexDirection;
-  flexWrap?: FlexWrap;
-  justifyContent?: JustifyContent;
-  alignItems?: AlignItems;
-  borderRadius?: BorderRadius;
-  gap?: Gap;
-  padding?: Padding;
-  paddingLeft?: PaddingLeft;
-  paddingRight?: PaddingRight;
-  paddingTop?: PaddingTop;
-  paddingBottom?: PaddingBottom;
-  paddingX?: PaddingX;
-  paddingY?: PaddingY;
-  margin?: Margin;
-  marginLeft?: MarginLeft;
-  marginRight?: MarginRight;
-  marginTop?: MarginTop;
-  marginBottom?: MarginBottom;
-  marginX?: MarginX;
-  marginY?: MarginY;
+  display?: DisplayProps;
+  flexDirection?: FlexDirectionProps;
+  flexWrap?: FlexWrapProps;
+  justifyContent?: JustifyContentProps;
+  alignItems?: AlignItemsProps;
+  borderRadius?: BorderRadiusProps;
   className?: string;
   style?: React.CSSProperties;
 }
