+++ packages/canon/src/components/Stack/styles.css
@@ -1,0 +1,4 @@
+.stack {
+  display: flex;
+  flex-direction: column;
+}
