+++ packages/canon/src/layout/properties.ts
