+++ packages/canon/src/theme/styles.css
@@ -23,3 +23,4 @@
 
 /* Components */
 @import '../components/Button/styles.css';
+@import '../components/Stack/styles.css';
